Added some endpoints
