--- src/main/java/az/ac/cput/Main.java
@@ -1,7 +1,0 @@
-package az.ac.cput;
-
-public class Main {
-    public static void main(String[] args) {
-        System.out.println("Hello world!");
-    }
-}
