--- src/main/java/az/ac/cput/backend/BackendApplication.java
@@ -1,18 +1,0 @@
-package az.ac.cput.backend;
-/**
- *  BackendApplication.java
- *  This is the main class for the backend
- * Author: Peter Buckingham (220165289)
- * Date: 05 April 2023
- */
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BackendApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(BackendApplication.class, args);
-    }
-
-}
