--- src/main/java/az/ac/cput/repository/IInsuranceRepository.java
@@ -1,11 +1,0 @@
-package az.ac.cput.repository;
-
-import az.ac.cput.domain.Insurance;
-
-import java.util.List;
-
-public interface IInsuranceRepository extends IRepository <Insurance, Integer>{
-    List<Insurance> getAllInsurancePolicies();
-    Insurance getInsuranceById(Integer id);
-
-}
