--- src/main/java/az/ac/cput/scratch/Car.java
@@ -1,16 +1,0 @@
-package az.ac.cput.scratch;
-/**
- * Scratch Concepts:
- * Car.java
- * Class for the Car
- * Author: Peter Buckingham (220165289)
- * Date: 29 March 2023
- */
-
-public class Car {
-    private String make;
-    private String model;
-    private int year;
-    private String licensePlate;
-    // constructors, getters, setters
-}
