+++ src/main/java/za/ac/cput/Main.java
@@ -1,0 +1,7 @@
+package za.ac.cput;
+
+public class Main {
+    public static void main(String[] args) {
+        System.out.println("Hello world!");
+    }
+}
