+++ src/main/java/za/ac/cput/backend/controllers/GreetingController.java
@@ -1,0 +1,18 @@
+package za.ac.cput.backend.controllers;
+import java.util.concurrent.atomic.AtomicLong;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import za.ac.cput.backend.restservice.Greeting;
+
+@RestController
+public class GreetingController {
+    private static final String template = "Hello, %s!";
+    private final AtomicLong counter = new AtomicLong();
+
+    @GetMapping("/greeting")
+    public Greeting greeting(@RequestParam(value = "name", defaultValue = "World") String name) {
+        return new Greeting(counter.incrementAndGet(), String.format(template, name));
+    }
+}
