+++ src/main/java/za/ac/cput/backend/restservice/Greeting.java
@@ -1,0 +1,5 @@
+package za.ac.cput.backend.restservice;
+
+
+
+public record Greeting(long id, String content) { }
