+++ src/main/java/za/ac/cput/domain/IVehicle.java
@@ -1,0 +1,18 @@
+package za.ac.cput.domain;
+/**
+ * Vehicle.java
+ * Interface for the Vehicle
+ * Author: Peter Buckingham (220165289)
+ * Date: 29 March 2021
+ */
+
+public interface IVehicle {
+    int getId();
+    String getMake();
+    String getModel();
+    int getYear();
+    String getCategory();
+    String getLicensePlate();
+    boolean equals(Object o) ;
+    public int hashCode() ;
+}
