+++ src/main/java/za/ac/cput/infrastructure/IDatabaseConnector.java
@@ -1,0 +1,16 @@
+package za.ac.cput.infrastructure;
+/**
+ * IDatabaseConnector.java
+ * Interface for the DatabaseConnector
+ * Author: Peter Buckingham (220165289)
+ * Date: 19 March 2021
+ */
+
+import java.sql.Connection;
+
+
+public interface IDatabaseConnector {
+    Connection getConnection();
+
+    void closeConnection();
+}
