--- src/main/java/za/ac/cput/onthegorentalsspringboot/OnTheGoRentalsSpringBootApplication.java
@@ -1,13 +1,0 @@
-package za.ac.cput.onthegorentalsspringboot;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class OnTheGoRentalsSpringBootApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(OnTheGoRentalsSpringBootApplication.class, args);
-    }
-
-}
