+++ src/main/java/za/ac/cput/repository/IInsuranceRepository.java
@@ -1,0 +1,11 @@
+package za.ac.cput.repository;
+
+import za.ac.cput.domain.Insurance;
+
+import java.util.List;
+
+public interface IInsuranceRepository extends IRepository <Insurance, Integer>{
+    List<Insurance> getAllInsurancePolicies();
+    Insurance getInsuranceById(Integer id);
+
+}
