+++ src/main/java/za/ac/cput/scratch/Customer.java
@@ -1,0 +1,15 @@
+package za.ac.cput.scratch;
+/**
+ * Scratch Concepts:
+ * Customer.java
+ * Class for the Customer
+ * Author: Peter Buckingham (220165289)
+ * Date: 29 March 2023
+ */
+
+public class Customer {
+    private String firstName;
+    private String lastName;
+    private String email;
+    // constructors, getters, setters
+}
