+++ src/main/java/za/ac/cput/scratch/Rental.java
@@ -1,0 +1,17 @@
+package za.ac.cput.scratch;
+
+/**
+ * Scratch Concepts:
+ * Rental.java
+ * Class for the Rental
+ * Author: Peter Buckingham (220165289)
+ * Date: 29 March 2023
+ */
+import java.util.Date;
+
+public class Rental {
+    private Car car;
+    private Date startDate;
+    private Date endDate;
+    // constructors, getters, setters
+}
