+++ src/main/java/za/ac/cput/scratch/Reservation.java
@@ -1,0 +1,13 @@
+package za.ac.cput.scratch;
+/**
+ * scratch concepts
+ * Reservation.java
+ * Class for the Reservation
+ * Author: Peter Buckingham (220165289)
+ * Date: 29 March 2023
+ */
+public class Reservation {
+    private Customer customer;
+    private Rental rental;
+    // constructors, getters, setters
+}
